renamed itr skippable to TIA
